Clean up type coercions
git-svn-id: http://svn.code.sf.net/p/netpbm/code/trunk@4321 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ lib/pmfileio.c
@@ -633,8 +633,8 @@ pm_readbigshort(FILE *  const ifP,
 
     unsigned short s;
 
-    s  = getcNofail(ifP) << 8;
-    s |= getcNofail(ifP) << 0;
+    s  = (unsigned short)getcNofail(ifP) << 8;
+    s |= (unsigned short)getcNofail(ifP) << 0;
 
     *sP = s;
 
@@ -661,10 +661,10 @@ pm_readbiglong(FILE * const ifP,
 
     unsigned long l;
 
-    l  = getcNofail(ifP) << 24;
-    l |= getcNofail(ifP) << 16;
-    l |= getcNofail(ifP) <<  8;
-    l |= getcNofail(ifP) <<  0;
+    l  = (unsigned long)getcNofail(ifP) << 24ul;
+    l |= (unsigned long)getcNofail(ifP) << 16ul;
+    l |= (unsigned long)getcNofail(ifP) <<  8ul;
+    l |= (unsigned long)getcNofail(ifP) <<  0ul;
 
     *lP = l;
 
@@ -709,8 +709,8 @@ pm_readlittleshort(FILE *  const ifP,
                    short * const sP) {
     unsigned short s;
 
-    s  = getcNofail(ifP) << 0;
-    s |= getcNofail(ifP) << 8;
+    s  = (unsigned short)getcNofail(ifP) << 0;
+    s |= (unsigned short)getcNofail(ifP) << 8;
 
     *sP = s;
 
@@ -736,10 +736,10 @@ pm_readlittlelong(FILE * const ifP,
                   long * const lP) {
     unsigned long l;
 
-    l  = getcNofail(ifP) <<  0;
-    l |= getcNofail(ifP) <<  8;
-    l |= getcNofail(ifP) << 16;
-    l |= getcNofail(ifP) << 24;
+    l  = (unsigned long)getcNofail(ifP) <<  0ul;
+    l |= (unsigned long)getcNofail(ifP) <<  8ul;
+    l |= (unsigned long)getcNofail(ifP) << 16ul;
+    l |= (unsigned long)getcNofail(ifP) << 24ul;
 
     *lP = l;
